+++ CheapHelpers.Blazor/CheapHelpers.Blazor.csproj
@@ -13,6 +13,7 @@
 
   <ItemGroup>
     <PackageReference Include="BlazorDownloadFile" Version="2.4.0.2" />
+    <PackageReference Include="MimeMapping" Version="3.1.0" />
     <PackageReference Include="MudBlazor" Version="8.6.0" />
   </ItemGroup>
