+++ CheapHelpers.Blazor/Helpers/Extensions.cs
@@ -1,4 +1,5 @@
 ﻿using CheapHelpers.EF;
+using CheapHelpers.Models;
 using Microsoft.AspNetCore.Components;
 using MudBlazor;
 using System;
