+++ CheapHelpers.EF/ContextExtensions.cs
@@ -5,7 +5,7 @@ using System.Diagnostics;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using CheapHelpers;
+using CheapHelpers.Models;
 
 namespace CheapHelpers.EF
 {
