+++ CheapHelpers.Models/Models/IDisplayIndex.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace CheapHelpers
+namespace CheapHelpers.Models
 {
     /// <summary>
     /// Move DisplayIndex recalculations te here or atleast use the interface
