+++ CheapHelpers.Models/Models/IEntityId.cs
@@ -1,4 +1,4 @@
-﻿namespace CheapHelpers
+﻿namespace CheapHelpers.Models
 {
     public interface IEntityId
     {
