+++ CheapHelpers.Services/Services/BarcodeService.cs
