+++ CheapHelpers.Services/Services/IBarcodeService.cs
