+++ CheapHelpers.Services/Services/IXmlService.cs
