+++ CheapHelpers.Services/Services/XmlService.cs
