+++ CheapHelpers.Services/WebServices/IWebServiceBase.cs
