+++ CheapHelpers.sln
@@ -1,13 +1,17 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.14.36109.1 d17.14
+VisualStudioVersion = 17.14.36109.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CheapHelpers.Blazor", "CheapHelpers.Blazor\CheapHelpers.Blazor.csproj", "{3DC2B4B3-62DB-4601-B483-2A6B7E14F2D1}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CheapHelpers.EF", "CheapHelpers.EF\CheapHelpers.EF.csproj", "{ABDEE8A6-9E24-4066-B5BF-B9B85666AA66}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CheapHelpers", "CheapHelpers\CheapHelpers.csproj", "{57181932-7A2F-5D28-92F6-1B6CED552474}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CheapHelpers.Services", "CheapHelpers.Services\CheapHelpers.Services.csproj", "{357BDD2C-A54D-402F-8E53-87158C85B2B9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CheapHelpers.Models", "CheapHelpers.Models\CheapHelpers.Models.csproj", "{8D55B555-C4F4-4577-B75C-3CC056A30FA0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +31,14 @@ Global
 		{57181932-7A2F-5D28-92F6-1B6CED552474}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{57181932-7A2F-5D28-92F6-1B6CED552474}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{57181932-7A2F-5D28-92F6-1B6CED552474}.Release|Any CPU.Build.0 = Release|Any CPU
+		{357BDD2C-A54D-402F-8E53-87158C85B2B9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{357BDD2C-A54D-402F-8E53-87158C85B2B9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{357BDD2C-A54D-402F-8E53-87158C85B2B9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{357BDD2C-A54D-402F-8E53-87158C85B2B9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8D55B555-C4F4-4577-B75C-3CC056A30FA0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8D55B555-C4F4-4577-B75C-3CC056A30FA0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8D55B555-C4F4-4577-B75C-3CC056A30FA0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8D55B555-C4F4-4577-B75C-3CC056A30FA0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
